% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{package}
\name{townsendr-package}
\alias{townsendr}
\alias{townsendr-package}
\title{townsendr: Calculate Townsend Material Deprivation Scores}
\description{
Calculate Townsend Material Deprivation scores for various geographies and censuses.
}
\author{
\strong{Maintainer}: Phil Mike Jones \email{philmikejones@gmail.com}

}
\keyword{internal}
