% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_calc_townsend.R
\name{tr_calc_townsend}
\alias{tr_calc_townsend}
\title{tr_calc_townsend}
\usage{
tr_calc_townsend(year = NULL, geography = NULL)
}
\arguments{
\item{year}{a census year (2011, 2001, 1991, 1981)}

\item{geography}{the geography you wish to calculate Townsend scores for}
}
\description{
Calculate Townsend Material Deprivation scores for the specified geography.
}
