% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_list_geographies.R
\name{tr_list_geographies}
\alias{tr_list_geographies}
\title{tr_list_geographies}
\usage{
tr_list_geographies(year = NULL)
}
\arguments{
\item{year}{A census year}
}
\value{
A [tibble][tibble::tibble-package] of available census geographies
}
\description{
List the available census geographies. Returns a
[tibble][tibble::tibble-package] with one row per available geography.
}
\examples{
\dontrun{
tr_list_geographies(2011)
}
}
