% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_calc_townsend.R
\name{tr_calc_townsend}
\alias{tr_calc_townsend}
\title{tr_calc_townsend}
\usage{
tr_calc_townsend(year = NULL, geography = NULL)
}
\arguments{
\item{year}{a census year (2011, 2001, 1991, 1981)}

\item{geography}{the geography you wish to calculate Townsend scores for}
}
\value{
A tibble with census year, geography code, geography name, and results
}
\description{
Calculate Townsend Material Deprivation scores for the specified geography.
Geographies are specified using the Nomis API format, and can be found using
tr_list_geographies()
}
\examples{
tr_calc_townsend(2011, "TYPE480")
}
