% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_check_year.R
\name{tr_check_year}
\alias{tr_check_year}
\title{tr_check_year}
\usage{
tr_check_year(year)
}
\arguments{
\item{year}{an input that can be converted to an integer year}
}
\value{
Returns silently or, if an incorrect year is specified, an error
}
\description{
Check a census year input
}
\examples{
tr_check_year(2011)  # returns silently
tr_check_year(2012)  # returns an error
}
