% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_list_geographies.R
\name{tr_list_geographies}
\alias{tr_list_geographies}
\title{tr_list_geographies}
\usage{
tr_list_geographies(year)
}
\arguments{
\item{year}{A census year}
}
\value{
A list of tibbles of available census geographies
}
\description{
tr_list_geographies
}
\examples{
tr_list_geographies(2011)
}
